requirements.txt
python-telegram-bot[job-queue]==22.6
supabase==2.4.3
python-dotenv==1.0.1
APScheduler==3.10.4
pytz==2024.1
flask==3.1.0
gunicorn==23.0.0
